+++ search/searcher/util_knn.go
@@ -18,9 +18,11 @@
 package searcher
 
 import (
+	"context"
 	"math"
 
 	"github.com/blevesearch/bleve/v2/search"
+	index "github.com/blevesearch/bleve_index_api"
 )
 
 // util func used by both disjunction and conjunction searchers
@@ -75,4 +77,46 @@ func (s *DisjunctionHeapSearcher) computeQueryNorm() {
 
 func (s *ConjunctionSearcher) computeQueryNorm() {
 	s.queryNorm, s.queryNormForKNN = computeQueryNorm(s.searchers)
+}
+
+func optimizeCompositeSearcher(ctx context.Context, optimizationKind string,
+	indexReader index.IndexReader, qsearchers []search.Searcher,
+	options search.SearcherOptions) (search.Searcher, error) {
+	var octx index.OptimizableContext
+
+	for _, searcher := range qsearchers {
+		// if is KNN searcher, continue
+		// should not break due to a kNN searcher.
+		if _, ok := searcher.(*KNNSearcher); ok {
+			continue
+		}
+
+		o, ok := searcher.(index.Optimizable)
+		if !ok {
+			return nil, nil
+		}
+
+		var err error
+		octx, err = o.Optimize(optimizationKind, octx)
+		if err != nil {
+			return nil, err
+		}
+
+		if octx == nil {
+			return nil, nil
+		}
+	}
+
+	optimized, err := octx.Finish()
+	if err != nil || optimized == nil {
+		return nil, err
+	}
+
+	tfr, ok := optimized.(index.TermFieldReader)
+	if !ok {
+		return nil, nil
+	}
+
+	return newTermSearcherFromReader(indexReader, tfr,
+		[]byte(optimizationKind), "*", 1.0, options)
 }
